consumer-configs/requirements.txt
pyspark==3.0.2
kafka-python
pymongo==4.4.0